Put organ and tissue classes in a separate file
