+++ Models/Models.csproj
@@ -119,6 +119,7 @@
   <ItemGroup>
     <Compile Include="AgPastureProto.cs" />
     <Compile Include="AgPasture\AgPasture.PastureSpecies.cs" />
+    <Compile Include="AgPasture\AgPasture.PastureSpecies.Organs.cs" />
     <Compile Include="AgPasture\AgPasture.SwardProto.cs" />
     <Compile Include="AgPasture\AgPasture.PastureSpeciesSwd.cs" />
     <Compile Include="AgPasture\AgPasture.Sward.cs" />
